+++ src/type-storage/type_storage.hpp
@@ -6,30 +6,39 @@
 #pragma once
 
 #include <tuple>
+#include <type_traits>
+#include <utility>
 
 namespace type_storage
 {
     namespace detail
     {
+
         /// Helper struct for find_index. "Empty" non specialised struct.
-        template<bool cond, typename... Ts>
+        template<template<class, class> class P,
+                 int i, typename U, bool cond, typename... Ts>
         struct find_type
         {
-            static_assert(cond, "Type not found in provided typelist");
+            static_assert(i != 0, "Type not found in provided typelist");
+            static_assert(1 >= i , "Ambiguous types in provided typelist");
+            static_assert(sizeof...(Ts) == 0, "Internal error: Ts not 0");
         };
 
         /// Helper struct for find_index, true specialisation.
         /// template member index<>::value finds index based on the size of
-        /// remaining paramter pack Ts... when a type match has been found.
-        template<typename... Ts>
-        struct find_type<true, Ts...>
+        /// remaining parameter pack Ts... when a type match has been found.
+        template<template<class, class> class P,
+                 int i, typename U, typename... Ts>
+        struct find_type<P, i, U, true, Ts...>
+            : find_type<P, i+1, U, false, Ts...>
+
         {
             /// specifies the index of the found type
             template<typename... Types>
             struct index
             {
                 /// describes the index of requested type in Ts...
-                static const size_t value = sizeof...(Types) - sizeof...(Ts);
+                static const size_t value = sizeof...(Types) - sizeof...(Ts) -1;
             };
         };
 
@@ -37,40 +46,77 @@ namespace type_storage
         /// Check for type match between U and T, and inherits from correct
         /// find_type specialisation. member index<>::value obtained by
         /// recursion (recurses until U == T )
-        template<typename U, typename T, typename... Ts>
-        struct find_type<false, U, T, Ts...>
-            : find_type<std::is_same<U, T>::value, U, Ts...>
-        {};
-
+        template<template<class, class> class P,
+                 int i, typename U, typename T, typename... Ts>
+        struct find_type<P, i, U, false, T, Ts...>
+            : find_type<P, i, U, P<U, T>::value, Ts...>
+        { };
 
         /// Finds index of type U in template parameter pack Ts using find_type
-        template<typename U, typename... Ts>
+        /// with success predicate P<U, Ts>...
+        template<template<class, class> class P, class U, class... Ts>
         struct find_index
         {
             /// describes the index of U in Ts...
             static const size_t value =
-                find_type<false, U, Ts...>::template index <Ts...>::value;
+                find_type<P, 0, U, false, Ts...>::template index <Ts...>::value;
         };
     }
 
     /// Get a object of specific type T from tuple regardless of its position.
     /// This function is basically a simple implementation of 
     /// C++14 std::get<T>(std::tuple<Types>&) function, using C++11 code.
-    /// NB: If tuple contains more than one object of the specific type,
-    /// a reference to the first found object with type T is returned.
+    /// NB: If tuple contains more than one object of the reqeusted type T,
+    /// the code cannot compile. Multiple objects of type different from T may
+    /// exist in the tuple.
     /// If type T is not contained in tuple the code cannot compile.
     /// @param tup the tuple of objects to look in
     template<typename T, typename... Types>
-    T& get(std::tuple<Types...>& tup)
+    auto get(std::tuple<Types...>& tup)
+        -> decltype(std::get<detail::find_index<std::is_same, 
+                                                T, Types...>::value>(tup))
     {
-        return std::get<detail::find_index<T, Types...>::value>(tup);
+        return std::get<detail::find_index<std::is_same, 
+                                           T, Types...>::value>(tup);
     }
 
     // Const version of above
     template<typename T, typename... Types>
-    const T& get(const std::tuple<Types...>& tup)
+    auto get(const std::tuple<Types...>& tup)
+        -> decltype(std::get<detail::find_index<std::is_same,
+                                                T, Types...>::value>(tup))
     {
-        return std::get<detail::find_index<T, Types...>::value>(tup);
+        return std::get<detail::find_index<std::is_same,
+                                           T, Types...>::value>(tup);
+    }
+
+    /// Get an object from base type B from tuple regardless of its position.
+    /// This function is virtually identical to get(), however the matching
+    /// condition is on base types, not exact types.
+    /// NB: If tuple contains more than one object of a matching base type,
+    /// the code will not compile (static assertion). The same is valid if the
+    /// tuple does not contain any types with matching base.
+    /// Only the base type requested may cause a compile error, all other 
+    /// types may be identical or share the same base, as long as they are 
+    /// different from type B.
+    /// @param tup the tuple of objects to look in
+    template<typename B, typename... Types>
+    auto baget(std::tuple<Types...>& tup)
+        -> decltype(std::get<detail::find_index<std::is_base_of,
+                                                B, Types...>::value>(tup))
+    {
+        return std::get<detail::find_index<std::is_base_of,
+                                           B, Types...>::value>(tup);
+    }
+
+    // Const version of above
+    template<typename B, typename... Types>
+    auto baget(const std::tuple<Types...>& tup)
+        -> decltype(std::get<detail::find_index<std::is_base_of,
+                                                B, Types...>::value>(tup))
+    {
+        return std::get<detail::find_index<std::is_base_of,
+                                           Types...>::value>(tup);
     }
 }