+++ NEWS.rst
@@ -6,7 +6,7 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Added baget function to get an object based on a specified base type.
 
 2.0.2
 -----